promptflow/document-data-extraction/requirements.txt
promptflow
azure-identity
azure-storage-blob